moving profiler into main.go
